pyproject.toml
[tool.poetry]
name = "shb2bot"
version = "0.1.0"
description = ""
authors = ["sysnote8main <sysnote8main@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
